feat(recipes): prevent deleting used recipes

+++ MealBuilder.Web/Pages/Recipes/Delete.cshtml.cs
@@ -2,6 +2,7 @@ using MealBuilder.Web.Data;
 using MealBuilder.Web.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
 
 namespace MealBuilder.Web.Pages.Recipes
 {
@@ -16,6 +17,8 @@ namespace MealBuilder.Web.Pages.Recipes
 
         public Recipe Recipe { get; set; } = new();
 
+        public bool IsUsed { get; set; }
+
         public async Task<IActionResult> OnGetAsync(int id)
         {
             Recipe? recipe = await _context.Recipes.FindAsync(id);
@@ -26,6 +29,7 @@ namespace MealBuilder.Web.Pages.Recipes
             }
 
             Recipe = recipe;
+            IsUsed = await IsRecipeUsedAsync(id);
 
             return Page();
         }
@@ -39,10 +43,29 @@ namespace MealBuilder.Web.Pages.Recipes
                 return NotFound();
             }
 
+            if (await IsRecipeUsedAsync(id))
+            {
+                Recipe = recipe;
+                IsUsed = true;
+
+                return Page();
+            }
+
             _context.Recipes.Remove(recipe);
             await _context.SaveChangesAsync();
 
             return RedirectToPage("./Index");
+        }
+
+        private async Task<bool> IsRecipeUsedAsync(int recipeId)
+        {
+            return await _context.RecipeComponents.AnyAsync(recipeComponent =>
+                       recipeComponent.ParentRecipeId == recipeId
+                       || recipeComponent.ComponentRecipeId == recipeId)
+                   || await _context.MenuItems.AnyAsync(menuItem =>
+                       menuItem.RecipeId == recipeId)
+                   || await _context.PreparedRecipeBatches.AnyAsync(preparedRecipeBatch =>
+                       preparedRecipeBatch.RecipeId == recipeId);
         }
     }
 }
